created customer register + developing sale + select product

+++ data/src/main/java/com/wisnitech/omiesales/data/model/Sale.kt
@@ -19,7 +19,7 @@ import kotlinx.parcelize.Parcelize
 )
 data class Sale(
     @PrimaryKey(autoGenerate = true)
-    val id: Int = 0,
-    val customerId: Int,
+    val id: Long = 0,
+    val customerId: Long,
     val saleDate: String
 ) : Parcelable
