+++ data/src/main/java/com/wisnitech/omiesales/data/model/SaleProduct.kt
@@ -21,7 +21,7 @@ import androidx.room.ForeignKey
     ]
 )
 data class SaleProduct(
-    val saleId: Int,
-    val productId: Int,
+    val saleId: Long,
+    val productId: Long,
     val quantity: Int
 )
