+++ data/src/main/java/com/wisnitech/omiesales/data/model/SumSales.kt
@@ -1,8 +1,8 @@
 package com.wisnitech.omiesales.data.model
 
 data class SumSales(
-    val saleId:Int,
-    val saleDate:String,
-    val customerName:String,
-    val saleValue:Double
+    val saleId: Long,
+    val saleDate: String,
+    val customerName: String,
+    val saleValue: Double
 )
